+++ app/common/types/urls.js
@@ -1,0 +1,7 @@
+const baseURL= 'http://localhost:9933';
+
+export default {
+  api: {
+    root: `${baseURL}/`,
+    }
+};